Refactored merge methods to be in the same service.
