+++ src/main/java/rs/teslaris/core/controller/ExpertiseOrSkillController.java
@@ -22,7 +22,6 @@ import rs.teslaris.core.dto.document.DocumentFileDTO;
 import rs.teslaris.core.dto.document.DocumentFileResponseDTO;
 import rs.teslaris.core.dto.person.ExpertiseOrSkillDTO;
 import rs.teslaris.core.dto.person.ExpertiseOrSkillResponseDTO;
-import rs.teslaris.core.dto.person.involvement.PersonCollectionEntitySwitchListDTO;
 import rs.teslaris.core.service.interfaces.person.ExpertiseOrSkillService;
 
 @Validated
@@ -84,15 +83,5 @@ public class ExpertiseOrSkillController {
     public void deleteExpertiseOrSkillProof(@PathVariable Integer expertiseOrSkillId,
                                             @PathVariable Integer proofId) {
         expertiseOrSkillService.deleteProof(proofId, expertiseOrSkillId);
-    }
-
-    @PatchMapping("/merge/person/source/{sourcePersonId}/target/{targetPersonId}")
-    @PreAuthorize("hasAuthority('MERGE_PERSON_METADATA')")
-    public void switchInvolvementsToOtherPerson(@PathVariable Integer sourcePersonId,
-                                                @PathVariable Integer targetPersonId,
-                                                @RequestBody
-                                                PersonCollectionEntitySwitchListDTO skillSwitchList) {
-        expertiseOrSkillService.switchSkills(skillSwitchList.getEntityIds(), sourcePersonId,
-            targetPersonId);
     }
 }
