+++ src/test/java/rs/teslaris/core/unit/MergeServiceTest.java
@@ -1,6 +1,8 @@
 package rs.teslaris.core.unit;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.atLeastOnce;
@@ -8,6 +10,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -30,10 +33,14 @@ import rs.teslaris.core.model.document.PersonDocumentContribution;
 import rs.teslaris.core.model.document.Proceedings;
 import rs.teslaris.core.model.document.ProceedingsPublication;
 import rs.teslaris.core.model.institution.OrganisationUnit;
+import rs.teslaris.core.model.person.Education;
 import rs.teslaris.core.model.person.Employment;
+import rs.teslaris.core.model.person.ExpertiseOrSkill;
+import rs.teslaris.core.model.person.Involvement;
 import rs.teslaris.core.model.person.InvolvementType;
 import rs.teslaris.core.model.person.Person;
 import rs.teslaris.core.model.person.PersonName;
+import rs.teslaris.core.model.person.Prize;
 import rs.teslaris.core.repository.document.DocumentRepository;
 import rs.teslaris.core.repository.document.JournalPublicationRepository;
 import rs.teslaris.core.repository.document.ProceedingsPublicationRepository;
@@ -44,8 +51,11 @@ import rs.teslaris.core.service.interfaces.document.JournalPublicationService;
 import rs.teslaris.core.service.interfaces.document.JournalService;
 import rs.teslaris.core.service.interfaces.document.ProceedingsPublicationService;
 import rs.teslaris.core.service.interfaces.document.ProceedingsService;
+import rs.teslaris.core.service.interfaces.person.ExpertiseOrSkillService;
+import rs.teslaris.core.service.interfaces.person.InvolvementService;
 import rs.teslaris.core.service.interfaces.person.OrganisationUnitService;
 import rs.teslaris.core.service.interfaces.person.PersonService;
+import rs.teslaris.core.service.interfaces.person.PrizeService;
 import rs.teslaris.core.service.interfaces.user.UserService;
 
 @SpringBootTest
@@ -89,6 +99,15 @@ public class MergeServiceTest {
 
     @Mock
     private ProceedingsPublicationRepository proceedingsPublicationRepository;
+
+    @Mock
+    private PrizeService prizeService;
+
+    @Mock
+    private ExpertiseOrSkillService expertiseOrSkillService;
+
+    @Mock
+    private InvolvementService involvementService;
 
     @InjectMocks
     private MergeServiceImpl mergeService;
@@ -386,5 +405,143 @@ public class MergeServiceTest {
         verify(proceedingsPublicationRepository, times(2)).save(any(ProceedingsPublication.class));
         assertEquals(targetProceedings, publication1.getProceedings());
         assertEquals(targetProceedings, publication2.getProceedings());
+    }
+
+    @Test
+    void testSwitchInvolvements() {
+        // Given
+        var sourcePersonId = 1;
+        var targetPersonId = 2;
+        var involvementIds = List.of(100, 101, 102);
+
+        var sourcePerson = new Person();
+        sourcePerson.setId(sourcePersonId);
+        var sourceInvolvements = new HashSet<Involvement>();
+        involvementIds.forEach(id -> {
+            var involvement = new Involvement();
+            involvement.setId(id);
+            involvement.setPersonInvolved(sourcePerson);
+            sourceInvolvements.add(involvement);
+        });
+        sourcePerson.setInvolvements(sourceInvolvements);
+
+        var targetPerson = new Person();
+        targetPerson.setId(targetPersonId);
+        var targetInvolvements = new HashSet<Involvement>();
+        targetPerson.setInvolvements(targetInvolvements);
+
+        when(personService.findOne(sourcePersonId)).thenReturn(sourcePerson);
+        when(personService.findOne(targetPersonId)).thenReturn(targetPerson);
+        involvementIds.forEach(id -> {
+            var involvement = new Education();
+            involvement.setId(id);
+            when(involvementService.findOne(id)).thenReturn(involvement);
+        });
+
+        // When
+        mergeService.switchInvolvements(involvementIds, sourcePersonId, targetPersonId);
+
+        // Then
+        involvementIds.forEach(id -> {
+            var involvement = new Involvement();
+            involvement.setId(id);
+            verify(involvementService, times(1)).save(any(Involvement.class));
+        });
+        verify(personService).save(sourcePerson);
+        verify(personService).save(targetPerson);
+        verify(userService).updateResearcherCurrentOrganisationUnitIfBound(sourcePersonId);
+        verify(userService).updateResearcherCurrentOrganisationUnitIfBound(targetPersonId);
+        verify(personService).indexPerson(sourcePerson, sourcePersonId);
+        verify(personService).indexPerson(targetPerson, targetPersonId);
+    }
+
+    @Test
+    void testSwitchSkills() {
+        // Given
+        var sourcePersonId = 1;
+        var targetPersonId = 2;
+        var skillIds = List.of(100, 101, 102);
+
+        var sourcePerson = new Person();
+        sourcePerson.setId(sourcePersonId);
+        var sourceSkills = new HashSet<ExpertiseOrSkill>();
+        skillIds.forEach(id -> {
+            var skill = new ExpertiseOrSkill();
+            skill.setId(id);
+            sourceSkills.add(skill);
+        });
+        sourcePerson.setExpertisesAndSkills(sourceSkills);
+
+        var targetPerson = new Person();
+        targetPerson.setId(targetPersonId);
+        var targetSkills = new HashSet<ExpertiseOrSkill>();
+        targetPerson.setExpertisesAndSkills(targetSkills);
+
+        when(personService.findOne(sourcePersonId)).thenReturn(sourcePerson);
+        when(personService.findOne(targetPersonId)).thenReturn(targetPerson);
+        skillIds.forEach(id -> {
+            var skill = new ExpertiseOrSkill();
+            skill.setId(id);
+            when(expertiseOrSkillService.findOne(id)).thenReturn(skill);
+        });
+
+        // When
+        mergeService.switchSkills(skillIds, sourcePersonId, targetPersonId);
+
+        // Then
+        skillIds.forEach(id -> {
+            var skill = new ExpertiseOrSkill();
+            skill.setId(id);
+            verify(expertiseOrSkillService).findOne(id);
+            assertFalse(sourcePerson.getExpertisesAndSkills().contains(skill));
+            assertTrue(targetPerson.getExpertisesAndSkills().contains(skill));
+        });
+        verify(personService).save(sourcePerson);
+        verify(personService).save(targetPerson);
+    }
+
+    @Test
+    void testSwitchPrizes() {
+        // Given
+        var sourcePersonId = 1;
+        var targetPersonId = 2;
+        var prizeIds = List.of(100, 101, 102);
+
+        var sourcePerson = new Person();
+        sourcePerson.setId(sourcePersonId);
+        var sourcePrizes = new HashSet<Prize>();
+        prizeIds.forEach(id -> {
+            var prize = new Prize();
+            prize.setId(id);
+            sourcePrizes.add(prize);
+        });
+        sourcePerson.setPrizes(sourcePrizes);
+
+        var targetPerson = new Person();
+        targetPerson.setId(targetPersonId);
+        var targetPrizes = new HashSet<Prize>();
+        targetPerson.setPrizes(targetPrizes);
+
+        when(personService.findOne(sourcePersonId)).thenReturn(sourcePerson);
+        when(personService.findOne(targetPersonId)).thenReturn(targetPerson);
+        prizeIds.forEach(id -> {
+            var prize = new Prize();
+            prize.setId(id);
+            when(prizeService.findOne(id)).thenReturn(prize);
+        });
+
+        // When
+        mergeService.switchPrizes(prizeIds, sourcePersonId, targetPersonId);
+
+        // Then
+        prizeIds.forEach(id -> {
+            var prize = new Prize();
+            prize.setId(id);
+            verify(prizeService).findOne(id);
+            assertFalse(sourcePerson.getPrizes().contains(prize));
+            assertTrue(targetPerson.getPrizes().contains(prize));
+        });
+        verify(personService).save(sourcePerson);
+        verify(personService).save(targetPerson);
     }
 }
